Create Braintree Credit Card Model
We want to persist braintree credit card attributes. The assumption
is that the payment source is immutable and the data associated
with the payment method token always remains the same. The
callers can explicitly call get_or_fetch! so that they are aware
that it would be storing stuff to the database.

Also get_or fetch is an explicit call rather than shoving it
inside an attribute so that the caller is aware that he might
be making a network fetch rather than just getting an attribute
on an object.